entropy2performance/verl/requirements.txt
# requirements.txt records the full set of dependencies for development
accelerate
codetiming
datasets
dill
flash-attn
hydra-core
liger-kernel
math-verify[antlr4_9_3]
numpy
pandas
peft
pyarrow>=15.0.0
pybind11
pylatexenc
ray[default]
tensordict<0.6
torchdata
transformers
wandb
